% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivalor_helpers.R
\name{bivalor_assert_inputs}
\alias{bivalor_assert_inputs}
\title{Validar entradas básicas para bivariados}
\usage{
bivalor_assert_inputs(edicion = NULL, ediciones = NULL, ind_x, ind_y)
}
\arguments{
\item{edicion}{integer(1) o NULL.}

\item{ediciones}{integer() o NULL.}

\item{ind_x, ind_y}{character(1). Código o nombre de indicador.}
}
\value{
invisible(TRUE) o error si falla.
}
\description{
Chequea que exista al menos una referencia temporal (edicion o ediciones),
y que ambos indicadores (X e Y) estén provistos.
}
