% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivalor_helpers.R
\name{bivalor_build_palette}
\alias{bivalor_build_palette}
\title{Paleta cualitativa para regiones (re-usa las de valor_*)}
\usage{
bivalor_build_palette(n, which = c("ipe", "okabe_ito", "viridis"))
}
\arguments{
\item{n}{integer(1).}

\item{which}{"ipe","okabe_ito","viridis".}
}
\value{
vector de colores.
}
\description{
Paleta cualitativa para regiones (re-usa las de valor_*)
}
