% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogos.R
\name{catalogo_anio}
\alias{catalogo_anio}
\title{Cat치logo de a침os disponibles en INCORE}
\usage{
catalogo_anio()
}
\value{
tibble con columnas codigo y nombre
}
\description{
Cat치logo de a침os disponibles en INCORE
}
