% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogos.R
\name{catalogo_indicador}
\alias{catalogo_indicador}
\title{Cat치logo: Indicadores (c칩digo -> nombre exacto en la base)}
\usage{
catalogo_indicador()
}
\value{
tibble con columnas codigo, nombre
}
\description{
Cat치logo: Indicadores (c칩digo -> nombre exacto en la base)
}
