% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_grupo.R
\name{incorer_get_region_groups}
\alias{incorer_get_region_groups}
\title{Obtener los grupos de regiones definidos}
\usage{
incorer_get_region_groups()
}
\value{
named list con los grupos
}
\description{
Obtener los grupos de regiones definidos
}
