% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_grupo.R
\name{incorer_set_region_groups}
\alias{incorer_set_region_groups}
\title{Definir/actualizar grupos de regiones}
\usage{
incorer_set_region_groups(groups, replace = FALSE)
}
\arguments{
\item{groups}{named list. Cada elemento es un vector de nombres de regiones.}

\item{replace}{logical(1). Si TRUE, reemplaza todo; si FALSE (default), hace merge.}
}
\description{
Definir/actualizar grupos de regiones
}
