% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_consolidar_valor}
\alias{indc_consolidar_valor}
\title{Facilidad: asegurar un valor por (region, edicion, indicador)}
\usage{
indc_consolidar_valor(df)
}
\description{
Facilidad: asegurar un valor por (region, edicion, indicador)
}
\keyword{internal}
