% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_filtrar_regiones}
\alias{indc_filtrar_regiones}
\title{Facilidad: filtrar dataframe por regiones usando helper de grupos/códigos}
\usage{
indc_filtrar_regiones(df, regiones = "ALL", usar_codigos = TRUE)
}
\description{
Facilidad: filtrar dataframe por regiones usando helper de grupos/códigos
}
\keyword{internal}
