% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_leer_incore}
\alias{indc_leer_incore}
\title{Leer base INCORE (alias interno para consistencia indc_)}
\usage{
indc_leer_incore(
  edicion,
  usar_codigos = TRUE,
  agregar_codigos = TRUE,
  verbose = FALSE
)
}
\description{
Leer base INCORE (alias interno para consistencia indc_)
}
\keyword{internal}
