% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_paleta_continua}
\alias{indc_paleta_continua}
\title{Paleta continua 0–10 para puntajes INCORE (helpers indc_)}
\usage{
indc_paleta_continua(
  paleta = c("blues", "greens", "viridis", "cividis", "magma", "divergente"),
  na_value = "grey90"
)
}
\description{
Paleta continua 0–10 para puntajes INCORE (helpers indc_)
}
\keyword{internal}
