% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_paleta_discreta}
\alias{indc_paleta_discreta}
\title{Paleta cualitativa por región (helpers indc_)}
\usage{
indc_paleta_discreta(n, which = c("ipe", "okabe_ito", "viridis"))
}
\description{
Paleta cualitativa por región (helpers indc_)
}
\keyword{internal}
