% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_pilares_canonicos}
\alias{indc_pilares_canonicos}
\title{Conjunto de nombres de pilares can贸nicos (nombre y c贸digo)}
\usage{
indc_pilares_canonicos(type = c("nombre", "codigo"))
}
\description{
Conjunto de nombres de pilares can贸nicos (nombre y c贸digo)
}
\keyword{internal}
