% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_resolver_indicador}
\alias{indc_resolver_indicador}
\title{Resolver indicador (código o nombre) → lista con código/nombre y pilar}
\usage{
indc_resolver_indicador(x, dic = NULL)
}
\value{
list(codigo, nombre, pilar_cod, pilar_nom, numero)
}
\description{
Resolver indicador (código o nombre) → lista con código/nombre y pilar
}
\details{
Tolera catálogos con columnas `cod`, `codigo` o `code`; y `pilar_cod`, `pilar`, etc.
}
\keyword{internal}
