% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_resolver_pilar}
\alias{indc_resolver_pilar}
\title{Resolver pilar (código o nombre) a nombre oficial}
\usage{
indc_resolver_pilar(pilar, usar_codigos = TRUE)
}
\description{
Resolver pilar (código o nombre) a nombre oficial
}
\keyword{internal}
