% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_resolver_regiones}
\alias{indc_resolver_regiones}
\title{Resolver regiones (grupos + exclusiones + códigos) para indc_}
\usage{
indc_resolver_regiones(regiones = "ALL", usar_codigos = TRUE)
}
\description{
Resolver regiones (grupos + exclusiones + códigos) para indc_
}
\keyword{internal}
