% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_traer_base_puntajes}
\alias{indc_traer_base_puntajes}
\title{Traer base de PUNTAJES (0–10) para indicadores (excluye 'General' por defecto)}
\usage{
indc_traer_base_puntajes(
  ediciones,
  usar_codigos = TRUE,
  incluir_peru = FALSE,
  incluir_general = FALSE,
  agregar_codigos = TRUE,
  verbose = FALSE
)
}
\description{
Traer base de PUNTAJES (0–10) para indicadores (excluye 'General' por defecto)
}
\keyword{internal}
