% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indc_helpers.R
\name{indc_validar_rango}
\alias{indc_validar_rango}
\title{Valida rango duro de INCORE para edicion/es (2016–2025)}
\usage{
indc_validar_rango(edicion = NULL, ediciones = NULL)
}
\description{
Valida rango duro de INCORE para edicion/es (2016–2025)
}
\keyword{internal}
