% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_grupo.R
\docType{data}
\name{region_grupos}
\alias{region_grupos}
\alias{.incorer_region_groups_env}
\title{Registrar y resolver grupos de regiones (uso global en el paquete)}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.incorer_region_groups_env
}
\description{
Registrar y resolver grupos de regiones (uso global en el paquete)
}
\keyword{internal}
