% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traducir_codigo.R
\name{traducir_codigo}
\alias{traducir_codigo}
\title{Traducir códigos a nombres oficiales usando un catálogo}
\usage{
traducir_codigo(x, dic)
}
\arguments{
\item{x}{vector de códigos o nombres}

\item{dic}{diccionario con columnas (codigo, nombre) o (code, label)}
}
\value{
vector de nombres traducidos; avisa si hay códigos no reconocidos
}
\description{
Traducir códigos a nombres oficiales usando un catálogo
}
