% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valor_helpers.R
\name{valor_formato_valor}
\alias{valor_formato_valor}
\title{Formateo numérico condicional por unidad}
\usage{
valor_formato_valor(x, unidad = NULL, accuracy = 0.01)
}
\arguments{
\item{x}{numeric()}

\item{unidad}{character(1) o `NULL`}

\item{accuracy}{numeric(1) para `scales::number()`}
}
\value{
character()
}
\description{
Formateo numérico condicional por unidad
}
