% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valor_helpers.R
\name{valor_fuente_caption}
\alias{valor_fuente_caption}
\title{Construir etiqueta de fuente global}
\usage{
valor_fuente_caption(df, preferir_ipe = TRUE, max_list = 3L)
}
\arguments{
\item{df}{data.frame con columna `fuente`}

\item{preferir_ipe}{logical(1)}

\item{max_list}{integer(1)}
}
\value{
character(1)
}
\description{
Construir etiqueta de fuente global
}
