% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valor_helpers.R
\name{valor_resolver_indicador}
\alias{valor_resolver_indicador}
\title{Resolver indicador (código o nombre) -> list(code, name)}
\usage{
valor_resolver_indicador(indicador = NULL, usar_codigos = TRUE)
}
\arguments{
\item{indicador}{character(1) o NULL}

\item{usar_codigos}{logical(1)}
}
\value{
list(code, name) (NULL/NULL si no se reconoce)
}
\description{
Resolver indicador (código o nombre) -> list(code, name)
}
