% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valor_helpers.R
\name{valor_resolver_pilar}
\alias{valor_resolver_pilar}
\title{Resolver pilar (código o nombre) -> nombre oficial}
\usage{
valor_resolver_pilar(pilar = NULL, usar_codigos = TRUE)
}
\arguments{
\item{pilar}{character(1) o NULL}

\item{usar_codigos}{logical(1)}
}
\value{
nombre del pilar (o el original si no se puede traducir)
}
\description{
Resolver pilar (código o nombre) -> nombre oficial
}
