% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valor_helpers.R
\name{valor_resolver_regiones}
\alias{valor_resolver_regiones}
\title{Resolver regiones (códigos o nombres o grupos) a nombres oficiales
Soporta:
  - "ALL"
  - tokens de grupo: "gr_costa", "gr_sierra", etc. (definidos por el usuario)
  - exclusiones con prefijo "-": p.ej. c("gr_costa","-Lima*")}
\usage{
valor_resolver_regiones(regiones = "ALL", usar_codigos = TRUE)
}
\description{
Resolver regiones (códigos o nombres o grupos) a nombres oficiales
Soporta:
  - "ALL"
  - tokens de grupo: "gr_costa", "gr_sierra", etc. (definidos por el usuario)
  - exclusiones con prefijo "-": p.ej. c("gr_costa","-Lima*")
}
